Change to work:)

+++ include/X10Bit.hpp
@@ -7,7 +7,7 @@ class X10Bit
 public:
 	X10Bit();
 	~X10Bit();
-	void write(int, int);
+	void write(uint64_t , uint8_t);
 	bool isSending();
 	//void simulate(Data bits);
 	virtual void onReceiveBit(bool bit);
